+++ EnhanceQoL/Locales/zhCN.lua
@@ -124,3 +124,5 @@ L["autoAcceptGroupInvite"] = "自动接受组队邀请"
 L["autoAcceptGroupInviteGuildOnly"] = "公会成员"
 L["autoAcceptGroupInviteFriendOnly"] = "好友"
 L["autoAcceptGroupInviteOptions"] = "接受来自...的邀请"
+
+L["showLeaderIconRaidFrame"] = "在团队风格的队伍框体上显示队长图标"